feat: search loading status added
