+++ pages/result/index.tsx
@@ -1,40 +1,21 @@
 import { useBeersQuery } from "@/../hooks/query/useBeerQuery";
-import { generateBeerDetailUrl } from "@/../utils/url";
-import { CommonBeerImage } from "@/components/shared/CommonBeerImage/CommonBeerImage";
-import {
-  Box,
-  Circle,
-  Container,
-  Flex,
-  HStack,
-  SimpleGrid,
-} from "@chakra-ui/react";
+import { Box, Circle, Container, Flex } from "@chakra-ui/react";
 import { useRouter } from "next/router";
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
+import { MIN_MAX_BEER_VOLUME_SLIDER } from "../../interface/static";
 import { CategoryFilterListType, CategoryTitle } from "../../interface/types";
 import { EmptyFilter, WhiteFilter } from "../../public/svg";
 import { SearchFilterList } from "../../src/components/result/filter/search-filter-list/SearchFilterList";
 import SearchInput from "../../src/components/search/SearchInput";
-import {
-  BeerCard,
-  BeerCardBody,
-  BeerCardFooter,
-  BeerCategoryTag,
-  BeerCategoryTagLabel,
-  BeerCountryText,
-  BeerNameText,
-} from "../../src/components/shared/Card/BeerCardItem";
 import { LeftBackTitle } from "../../src/components/shared/Headers/LeftBackTitle";
 import { BeerSortType } from "../../types/common";
-import { EmptyFilteredResult } from "@/components/result/EmptyFilteredResult";
-import { MIN_MAX_BEER_VOLUME_SLIDER } from "../../interface/static";
+import { SearchResult } from "./SearchResult";
 
 const SearchResultPage = () => {
   const router = useRouter();
   const { query } = router.query;
 
   const [isFilterListOpen, setIsFilterListOpen] = useState<boolean>(true);
-  const [value, setValue] = useState<string>("");
   // TODO: refactor filter into 1 object data
   const [selectedFilters, setSelectedFilter] = useState<
     CategoryFilterListType[]
@@ -44,9 +25,11 @@ const SearchResultPage = () => {
       title: CategoryTitle.SORT_CRITERIA,
     },
   ]);
+
   const _selectedSort = selectedFilters.find(
     (filter) => filter.title === CategoryTitle.SORT_CRITERIA
   )?.tags;
+
   const selectedSort = _selectedSort
     ? (_selectedSort[0] as BeerSortType)
     : BeerSortType.MOST_LIKES;
@@ -86,9 +69,6 @@ const SearchResultPage = () => {
     SearchBeerQuery.refetch();
   }, [selectedFilters, beerVolume]);
 
-  const clearValue = () => {
-    setValue("");
-  };
   const handleFocus = () => {
     router.push("/search");
   };
@@ -97,13 +77,6 @@ const SearchResultPage = () => {
     setIsFilterListOpen(!isFilterListOpen);
   };
 
-  const handleClickCard = useCallback(
-    (id: number, name: string) => {
-      const url = generateBeerDetailUrl(id, name);
-      router.push(url);
-    },
-    [router]
-  );
   const handleClickTag = (
     targetTitle: CategoryTitle,
     targetTag: string | number
@@ -172,11 +145,7 @@ const SearchResultPage = () => {
         <LeftBackTitle />
         <Box p={"68px 24px 24px"} h="full">
           <Flex gap="10px" alignItems="center" mb="10px" cursor={"pointer"}>
-            <SearchInput
-              clearValue={clearValue}
-              onFocus={handleFocus}
-              autoFocus={false}
-            />
+            <SearchInput onFocus={handleFocus} autoFocus={false} />
             <Circle size="31px" bg="blue.100" onClick={handleClickToggle}>
               {isFilterListOpen ? <WhiteFilter /> : <EmptyFilter />}
             </Circle>
@@ -191,49 +160,10 @@ const SearchResultPage = () => {
             onChangeBeerVolume={handleChangeBeerVolume}
           />
 
-          {SearchBeerQuery.data ? (
-            <SimpleGrid columns={2} spacing={"16px"} mt={"8px"}>
-              {SearchBeerQuery.data?.contents?.map((beerItems) => {
-                const { id, name, origin_country, image_url, category } =
-                  beerItems;
-                return (
-                  <BeerCard
-                    key={beerItems.id}
-                    mt={1}
-                    w="full"
-                    onClick={() => handleClickCard(id, name)}
-                  >
-                    <BeerCardBody w="full" h="full" position={"relative"}>
-                      <Box position="relative">
-                        {image_url && (
-                          <CommonBeerImage
-                            src={image_url}
-                            alt={name}
-                            width="175px"
-                            height="175px"
-                            objectFit="cover"
-                          />
-                        )}
-                      </Box>
-                    </BeerCardBody>
-                    <BeerCardFooter>
-                      <BeerNameText>{name}</BeerNameText>
-                      <HStack>
-                        <BeerCountryText country={origin_country} />
-                        <BeerCategoryTag>
-                          <BeerCategoryTagLabel>
-                            {category?.name}
-                          </BeerCategoryTagLabel>
-                        </BeerCategoryTag>
-                      </HStack>
-                    </BeerCardFooter>
-                  </BeerCard>
-                );
-              })}
-            </SimpleGrid>
-          ) : (
-            <EmptyFilteredResult />
-          )}
+          <SearchResult
+            loading={SearchBeerQuery.isLoading || SearchBeerQuery.isFetching}
+            beers={SearchBeerQuery.data?.contents}
+          />
           <Box h="64px" />
         </Box>
       </Container>
